🧬 Pull up the function loads template configuration to interface *TemplateConfigLoadable*.
🧬 Pull up the property *config_file_name* to abstract class *TemplateConfigLoadable*.
